update: 상품관리 기능구현중, 삭제시 confirm 알람띄우기

+++ src/views/user-admin/user-admin.js
@@ -9,14 +9,16 @@ import {
 
 const mainTag = document.getElementById("main__container");
 
-const adminPageList = document.querySelectorAll(
-  ".admin__page__list > button"
-);
-
 const adminBtnOder = document.querySelector(".btn__admin__oder");
 const adminBtnUser = document.querySelector(".btn__admin__user");
-const adminBtnCategory = document.querySelector(".btn__admin__addCategory");
-const adminBtnadddProduct = document.querySelector(".btn__admin__addProduct");
+const adminBtnCategory = document.querySelector(".btn__admin__category");
+const adminBtnaddProduct = document.querySelector(".btn__admin__product");
+const allBtns = [
+  adminBtnOder,
+  adminBtnUser,
+  adminBtnCategory,
+  adminBtnaddProduct,
+];
 
 const oderAdmin = [
   "주문관리",
@@ -51,14 +53,14 @@ function compareEnglish(lsName) {
   return "add__product";
 }
 
-for (let i = 0; i < adminPageList.length - 2; i++) {
-  const listBtn = adminPageList[i];
+for (let i = 0; i < allBtns.length; i++) {
+  const listBtn = allBtns[i];
   listBtn.addEventListener("click", (element) => {
     //지금 table이 있는지 확인하고 있다면 다 지우기
     const currentTable = document.querySelector(".bd-example");
     if (currentTable) currentTable.remove();
 
-    const listName = adminPageList[i].innerText;
+    const listName = allBtns[i].innerText;
     const listNameEn = compareEnglish(listName);
 
     const newHtml = document.createElement("div");
@@ -83,7 +85,9 @@ for (let i = 0; i < adminPageList.length - 2; i++) {
               shopStatus: data.shippingStatus,
             };
           });
-          return newDatas;
+          return newDatas.sort((a, b) => {
+            return new Date(b.date) - new Date(a.date);
+          });
         })
         .then((newDatas) => {
           newHtml.appendChild(createOderTable(oderAdmin, newDatas));
@@ -114,13 +118,16 @@ for (let i = 0; i < adminPageList.length - 2; i++) {
               role: data.role,
             };
           });
-          return newDatas;
+          return newDatas.sort((a, b) => {
+            return new Date(b.date) - new Date(a.date);
+          });
         })
         .then((newData) => {
           newHtml.appendChild(createUserTable(userAdmin, newData));
           mainTag.append(newHtml);
         })
         .then(() => {
+          productManagementCreate();
           userManagementEdit();
           userManagementDelete();
         });
@@ -142,14 +149,15 @@ for (let i = 0; i < adminPageList.length - 2; i++) {
               updateDate: data.updatedAt.slice(0, 10),
             };
           });
-          return newDatas;
+          return newDatas.sort((a, b) => {
+            return new Date(b.date) - new Date(a.date);
+          });
         })
         .then((newDatas) => {
           newHtml.appendChild(createCategoryTable(categoryAdmin, newDatas));
           mainTag.append(newHtml);
         })
         .then(() => {
-          categoryManagementCreate();
           categoryManagementEdit();
           categoryManagementDelete();
         });
@@ -173,20 +181,23 @@ for (let i = 0; i < adminPageList.length - 2; i++) {
               stock: data.stock,
             };
           });
-          return newDatas;
+          return newDatas.sort((a, b) => {
+            return new Date(b.date) - new Date(a.date);
+          });
         })
         .then((newDatas) => {
           newHtml.appendChild(createProductTable(productAdmin, newDatas));
           mainTag.append(newHtml);
         })
         .then(() => {
-          productManagementCreate();
           productManagementEdit();
           productManagementDelete();
         });
     }
   });
 }
+categoryManagementCreate();
+productManagementCreate();
 
 //============ 주문관련 =====================
 function oderManagementEdit() {
@@ -224,13 +235,16 @@ function oderManagementDelete() {
   const deleteBtns = document.querySelectorAll(".btn__delete");
   for (let count = 0; count < deleteBtns.length; count++) {
     deleteBtns[count].addEventListener("click", (e) => {
-      const btnId = e.target.parentElement.parentElement.id;
-      document.getElementById(`${btnId}`).remove();
-      fetch(`http://localhost:5000/api/orders/${btnId}`, {
-        method: "DELETE",
-      })
-        .then((res) => res.json())
-        .then((alt) => alert(alt));
+      const conf = confirm("해당 주문건을 삭제하시겠습니까?");
+      if (conf) {
+        const btnId = e.target.parentElement.parentElement.id;
+        document.getElementById(`${btnId}`).remove();
+        fetch(`http://localhost:5000/api/orders/${btnId}`, {
+          method: "DELETE",
+        })
+          .then((res) => res.json())
+          .then((alt) => alert(alt));
+      }
     });
   }
 }
@@ -269,13 +283,16 @@ function userManagementDelete() {
   const deleteBtns = document.querySelectorAll(".btn__delete");
   for (let count = 0; count < deleteBtns.length; count++) {
     deleteBtns[count].addEventListener("click", (e) => {
-      const btnId = e.target.parentElement.parentElement.id;
-      document.getElementById(`${btnId}`).remove();
-      fetch(`http://localhost:5000/api/users/${btnId}`, {
-        method: "DELETE",
-      })
-        .then((res) => res.json())
-        .then((idData) => alert(`${idData}이 삭제되었습니다.`));
+      const conf = confirm("해당 유저의 정보를 삭제하시겠습니까?");
+      if (conf) {
+        const btnId = e.target.parentElement.parentElement.id;
+        document.getElementById(`${btnId}`).remove();
+        fetch(`http://localhost:5000/api/users/${btnId}`, {
+          method: "DELETE",
+        })
+          .then((res) => res.json())
+          .then((idData) => alert(`${idData}이 삭제되었습니다.`));
+      }
     });
   }
 }
@@ -291,18 +308,16 @@ function categoryManagementEdit() {
       nameValue =
         document.querySelectorAll(".current__name")[count].innerText;
       productId = e.target.parentElement.parentElement.id;
-      const inputCategoryName = document.querySelector("#edit-category-name");
+      const inputCategoryName = document.getElementById("edit-category-name");
       inputCategoryName.value = nameValue;
-
-      //inputCategoryName.setAttribute("value", nameValue);
     });
   }
   //수정하기 버튼을 클릭했을 때
   document
     .querySelector(".submit__edit__category")
     .addEventListener("click", (e) => {
-      e.preventDefault();
-      const newValue = document.querySelector("#edit-category-name").value;
+      alert("클릭됐다");
+      const newValue = document.getElementById("edit-category-name").value;
       fetch(`http://localhost:5000/api/categories/${productId}`, {
         method: "PUT",
         headers: {
@@ -321,10 +336,7 @@ function categoryManagementEdit() {
         })
         .then((data) => {
           alert(`"${nameValue}"이(가) "${data.name}" 으로 변경되었습니다.`);
-          //기존에 있던 table 내의 카테고리 이름을 바뀐 카테고리 이름으로 바꾸어 그려줌
-          document
-            .getElementById(`${data._id}`)
-            .querySelector(".current__name").innerText = `${data.name}`;
+          document.querySelector(".btn__admin__category").click();
           bootstrap.Modal.getInstance("#btn__admin__editCategory").hide();
         })
         .catch(() => alert("카테고리 이름을 입력해주세요."));
@@ -370,75 +382,120 @@ function categoryManagementDelete() {
   const deleteBtns = document.querySelectorAll(".btn__delete");
   for (let count = 0; count < deleteBtns.length; count++) {
     deleteBtns[count].addEventListener("click", (e) => {
-      const btnId = e.target.parentElement.parentElement.id;
-      document.getElementById(`${btnId}`).remove();
-      fetch(`http://localhost:5000/api/categories/${btnId}`, {
-        method: "DELETE",
-      })
-        .then((res) => res.json())
-        .then((alt) => alert(alt));
+      const conf = confirm("해당 카테고리를 삭제하시겠습니까?");
+      if (conf) {
+        const btnId = e.target.parentElement.parentElement.id;
+        document.getElementById(`${btnId}`).remove();
+        fetch(`http://localhost:5000/api/categories/${btnId}`, {
+          method: "DELETE",
+        })
+          .then((res) => res.json())
+          .then((alt) => alert(alt));
+      }
     });
   }
 }
 //============== 상품관련 ===============
 function productManagementEdit() {
-  // const editProductBtns = document.querySelectorAll(".btn__admin__editProduct");
-  // let productId;
-  // let nameValue;
-  // let isCategories = false;
-  // for (let count = 0; count < editProductBtns.length; count++) {
-  //   editProductBtns[count].addEventListener("click", (e) => {
-  //     nameValue = document.querySelectorAll(".current__name")[count].innerText;
-  //     productId = e.target.parentElement.parentElement.id;
-  //     const inputCProductName = document.querySelector("#edit-product-name");
-  //     inputCProductName.value = nameValue;
-  //   });
-  // }
-  // //수정하기 버튼을 클릭했을 때
-  // document
-  //   .querySelector(".submit__edit__product")
-  //   .addEventListener("click", (e) => {
-  //     e.preventDefault();
-  //     const newValue = document.querySelector("#edit-product-name").value;
-  //     fetch(`http://localhost:5000/api/products/${productId}`, {
-  //       method: "PUT",
-  //       headers: {
-  //         "Content-Type": "application/json",
-  //       },
-  //       body: JSON.stringify({
-  //         name: `${newValue}`,
-  //       }),
-  //     })
-  //       .then((res) => {
-  //         return res.json();
-  //       })
-  //       .then((data) => {
-  //         alert(`"${nameValue}"이(가) "${data.name}" 으로 변경되었습니다.`);
-  //         //기존에 있던 table 내의 카테고리 이름을 바뀐 카테고리 이름으로 바꾸어 그려줌
-  //         document
-  //           .getElementById(`${data._id}`)
-  //           .querySelector(".current__name").innerText = `${data.name}`;
-  //         bootstrap.Modal.getInstance("#btn__admin__editProduct").hide();
-  //       });
-  //   });
+  //카테고리의 리스트를 불러오는 작업
+  const addCategoryList = document.querySelector(".btn__admin__editProduct");
+  addCategoryList.addEventListener("click", (e) => {
+    fetch("/api/categories")
+      .then((res) => res.json())
+      .then((datas) => {
+        document.getElementById("edit-product-category").innerHTML = "";
+        datas.forEach((element) => {
+          document.getElementById("edit-product-category").innerHTML += `
+      <option>${element.name}</option>`;
+        });
+      });
+  });
+
+  //수정하기 모달창을 띄울 때 데이터에 맞게 모달창에 넣어주기
+  const editProductBtns = document.querySelectorAll(
+    ".btn__admin__editProduct"
+  );
+  let productId;
+  let nameValue;
+
+  const name = document.getElementById("edit-product-name");
+  const category = document.getElementById("edit-product-category");
+  const shortDesc = document.getElementById("edit-short-description");
+  const longDesc = document.getElementById("edit-long-description");
+  const imageFile = document.getElementById("edit-product-img");
+  const stock = document.getElementById("edit-product-stock");
+  const price = document.getElementById("edit-product-price");
+
+  for (let count = 0; count < editProductBtns.length; count++) {
+    editProductBtns[count].addEventListener("click", (e) => {
+      productId = e.target.parentElement.parentElement.id;
+      fetch(`/api/products/${productId}`)
+        .then((res) => res.json())
+        .then((newData) => {
+          name.value = newData.name;
+          category.value = newData.category;
+          shortDesc.value = newData.shortDesc;
+          longDesc.value = newData.longDesc;
+          stock.value = newData.stock;
+          price.value = newData.price;
+        });
+    });
+  }
+  //수정하기 버튼을 클릭했을 때
+  const formData = new FormData();
+  if (category.value === "카테고리를 선택하세요") {
+    formData.append("category", "전체");
+  } else {
+    formData.append("category", category.value);
+  }
+  formData.append("name", name.value);
+  formData.append("shortDesc", shortDesc.value);
+  formData.append("longDesc", longDesc.value);
+  formData.append("productImage", imageFile.files[0]);
+  formData.append("stock", stock.value);
+  formData.append("price", price.value);
+
+  fetch("/api/products", {
+    method: "PUT",
+    body: formData,
+    headers: {},
+  })
+    .then(async (res) => {
+      const json = await res.json();
+      if (res.ok) {
+        return json;
+      }
+      return Promise.reject(json);
+    })
+    .then((data) => {
+      alert(`${data.name} 의 정보가 변경되었습니다.`);
+      //form 안의 input값 전부 초기화하기
+      name.value = "";
+      category.value = "";
+      shortDesc.value = "";
+      longDesc.value = "";
+      imageFile.value = "";
+      stock.value = "";
+      price.value = "";
+      bootstrap.Modal.getInstance("#btn__admin__editProduct").hide();
+      bootstrap.Modal.getInstance("#btn__admin__editProduct").hide();
+    })
+    .catch((err) => alert(err));
 }
 
 function productManagementCreate() {
   //카테고리의 리스트를 불러오는 작업
-  let isCategories = false;
   const addCategoryList = document.querySelector(".btn__admin__addProduct");
   addCategoryList.addEventListener("click", (e) => {
-    if (!isCategories) {
-      fetch("/api/categories")
-        .then((res) => res.json())
-        .then((datas) => {
-          datas.forEach((element) => {
-            document.getElementById("create-product-category").innerHTML += `
+    fetch("/api/categories")
+      .then((res) => res.json())
+      .then((datas) => {
+        document.getElementById("create-product-category").innerHTML = "";
+        datas.forEach((element) => {
+          document.getElementById("create-product-category").innerHTML += `
         <option>${element.name}</option>`;
-          });
         });
-      isCategories = true;
-    }
+      });
   });
   //추가하기 버튼을 클릭했을 때
   const addProductBtn = document.querySelector(".submit__product");
@@ -458,7 +515,6 @@ function productManagementCreate() {
       formData.append("category", category.value);
     }
     formData.append("name", name.value);
-    formData.append("category", category.value);
     formData.append("shortDesc", shortDesc.value);
     formData.append("longDesc", longDesc.value);
     formData.append("productImage", imageFile.files[0]);
@@ -489,7 +545,7 @@ function productManagementCreate() {
         price.value = "";
 
         bootstrap.Modal.getInstance("#btn__admin__addProduct").hide();
-        document.querySelector(".btn__admin__Product").click();
+        document.querySelector(".btn__admin__product").click();
       })
       .catch((err) => alert(err));
   });
@@ -498,6 +554,8 @@ function productManagementDelete() {
   const deleteBtns = document.querySelectorAll(".btn__delete");
   for (let count = 0; count < deleteBtns.length; count++) {
     deleteBtns[count].addEventListener("click", (e) => {
+      const conf = confirm("해당 상품의 정보를 삭제하시겠습니까?");
+      if (conf) {
       const btnId = e.target.parentElement.parentElement.id;
       document.getElementById(`${btnId}`).remove();
       fetch(`http://localhost:5000/api/products/${btnId}`, {
@@ -505,6 +563,7 @@ function productManagementDelete() {
       })
         .then((res) => res.json())
         .then((alt) => alert(alt));
+    }
     });
   }
 }
